Rename bpacking_simd_impl > bpacking_simd_kernel
